refactor(list-detail): extract VideoListContent component
Move kind-30005 video grid rendering out of ListDetailPage into a new
VideoListContent component. ListDetailPage is now thin: it fetches the
event, parses with parseVideoList, and delegates rendering to
VideoListContent. No behavior change for kind-30005; kind-30000 body is
left empty pending Task 6.5b.
